Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
psycopg2-binary
Flask-Migrate
